Renamed/refactored BoardParser
